+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-# Unreleases
+# Unreleased
+
+* Drop support for python 3.8 (\#438).
 * Update `_TaskBase` schema from fractal-common (\#431).
 
 # 1.0.5