+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Drop support for python 3.8 (\#438).
 * Update `_TaskBase` schema from fractal-common (\#431).
+* Update `fractal task edit` command (\#439).
 
 # 1.0.5